+++ example/settings.py
@@ -70,6 +70,14 @@ USE_TZ = True
 WSGI_APPLICATION = "example.wsgi.application"
 ASGI_APPLICATION = "example.asgi.application"
 
+CHANNEL_LAYERS = {
+    "default": {
+        "BACKEND": "channels_redis.core.RedisChannelLayer",
+        "CONFIG": {
+            "hosts": [("127.0.0.1", 6379)],
+        },
+    }
+}
 
 # Cache and database